backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
python-docx==1.1.0
python-pptx==0.6.23
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.2
email-validator==2.1.0
aiofiles==23.2.1
cors==0.1.0
gunicorn==21.2.0